+++ src/cli/__init__.py
@@ -1,4 +1,4 @@
-"""your_cli package."""
+"""cli package."""
 
 __all__ = ["__version__"]